requirement.txt
pytorch_lightning
torch
click
black
loguru
pyserini
faiss-cpu